% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazing_seq_dendro.R
\name{cluster.sequences}
\alias{cluster.sequences}
\title{Cluster sequences by calling Clustal Omega}
\usage{
cluster.sequences(seqs, clustalo.exe = "clustalo",
  out.distfile = tempfile(), out.fastafle = tempfile(),
  in.fastafile = tempfile(), ...)
}
\arguments{
\item{seqs}{either}

\item{...}{}
}
\description{
Cluster sequences by calling Clustal Omega
}
