% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazing_seq_dendro.R
\name{draw.msa.dendrogam}
\alias{draw.msa.dendrogam}
\title{Cluster and draw sequences}
\usage{
draw.msa.dendrogam(sequences, input.type = c("Fasta", "XStingSet"),
  distances = NULL, cutoff = 0.25, clustalo.exe = "clustalo")
}
\arguments{
\item{distances}{dist, input ditance matrix, if not providd the fasta file will be clustered and will be given internally}

\item{cutoff}{double, height to use as a cutoff for the hierarchial clustering}

\item{clustalo.exe}{sting, location of the custal omega executable or name of the command that calls it in your path}

\item{fastafile}{file, location of the fasta file}
}
\value{
ggplot object
}
\description{
Uses clustal omega to custer the sequeces prvided in a fata file and
Makes sequence logos of the cropped branches
}
\examples{


}
