% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amazing_seq_dendro.R
\name{get.seqs}
\alias{get.seqs}
\title{Get a biostrig object from different souces}
\usage{
get.seqs(seqs, ...)
}
\arguments{
\item{seqs}{loction of a fasta file or a character object in memory}

\item{...}{additonal arguments to be passed to other functions}
}
\value{
an XStringSet object
}
\description{
Get a biostrig object from different souces
}
